requirements.txt
streamlit
groq
pandas
openpyxl
python-docx
PyPDF2
language-tool-python
dotenv